requirements.txt
Flask==3.1.0
gdown==5.2.0
numpy==2.2.4
Pillow==11.1.0
tensorflow==2.19.0
gunicorn==23.0.0
python-dotenv==1.1.0